Organize types in model folder with barrel file
