+++ frontend/src/business/models/Api.ts
@@ -1,0 +1,7 @@
+export interface ApiState {
+  response: {
+    statusText?: string;
+    message: string;
+    status: "success" | "error" | undefined;
+  };
+}
