+++ frontend/src/business/models/App.ts
@@ -1,0 +1,13 @@
+export type NavItem = {
+  label: string;
+  subLabel?: string;
+  children?: Array<NavItem>;
+  href?: string;
+};
+
+export interface AppState {
+  isLoading: boolean;
+  publicNavItems: Array<NavItem>;
+  privateNavItems: Array<NavItem>;
+  colors: Array<string>;
+}
