+++ frontend/src/business/models/Auth.ts
@@ -1,0 +1,4 @@
+export interface AuthState {
+  isSignedIn: boolean;
+  token: string;
+}
