+++ frontend/src/business/models/index.ts
@@ -1,0 +1,13 @@
+export * from "./Api";
+export * from "./App";
+export * from "./Auth";
+export * from "./Author";
+export * from "./Book";
+export * from "./Bookworm";
+export * from "./Genre";
+export * from "./Language";
+export * from "./Rental";
+export * from "./Review";
+export * from "./State";
+export * from "./Table";
+export * from "./User";
