+++ frontend/src/business/overmind/App/state.ts
@@ -1,16 +1,4 @@
-export type NavItem = {
-  label: string;
-  subLabel?: string;
-  children?: Array<NavItem>;
-  href?: string;
-};
-
-export interface AppState {
-  isLoading: boolean;
-  publicNavItems: Array<NavItem>;
-  privateNavItems: Array<NavItem>;
-  colors: Array<string>;
-}
+import { AppState } from "../../models";
 
 export const state: AppState = {
   isLoading: false,
