+++ frontend/src/business/overmind/Genre/state.ts
@@ -1,12 +1,5 @@
 import { derived } from "overmind";
-import { BaseState } from "../../models/State";
-
-export interface Genre {
-  id: string;
-  genre: string;
-}
-
-export type GenreState = BaseState<Genre>;
+import { GenreState } from "../../models";
 
 export const state: GenreState = {
   getAllApi: null,
