+++ frontend/src/business/overmind/Review/state.ts
@@ -1,19 +1,6 @@
 import { derived } from "overmind";
 import { groupBy } from "../../functions";
-import { ApiResponse } from "../../models/State";
-
-export interface Review {
-  id: string;
-  description: string;
-  rating: number;
-  book_title: string;
-  reviewer: string;
-}
-
-export interface ReviewState {
-  getAllApi: ApiResponse<Array<Review>> | null;
-  overview: { [key: string]: Array<Review> } | null;
-}
+import { ReviewState } from "../../models";
 
 export const state: ReviewState = {
   getAllApi: null,
